backend/requirements/prod.txt
Django==5.1.7
psycopg2-binary==2.9.10
python-dotenv==1.0.1
djangorestframework==3.16.1
djoser==2.3.3
djangorestframework_simplejwt==5.5.1
celery==5.5.3
redis==6.4.0
flower==2.0.1
django-cors-headers==4.9.0
drf-yasg==1.21.11